// src/openaip/types/reporting_point.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

import { ISO31661Alpha2CountryCode } from "./country_code.js";

/**
* Response payload of a reporting point instance.
*/
export interface ReportingPointPayload {
 /**
  * The document's internal reference ID value.
  */
 _id?: string;
 name?: string;
 compulsory?: boolean;
 country?: ISO31661Alpha2CountryCode;
 geometry?: {
   type: "Point";
   /**
    * @minItems 2
    * @maxItems 2
    */
   coordinates: [number, number];
 };
 elevation?: {
   value: number;
   /**
    * The elevation unit. Always 'meters'.
    */
   unit: 0;
   /**
    * The elevation reference datum. Always 'MSL'.
    */
   referenceDatum?: 1;
 };
 elevationGeoid?: {
   /**
    * Height above ellipsoid in meters.
    */
   hae: number;
   /**
    * Height of geoid in meters.
    */
   geoidHeight: number;
 };
 airports?: string[];
 remarks?: string;
 /**
  * UID of user that created this document.
  */
 createdBy?: string;
 /**
  * UID of user that updated this document.
  */
 updatedBy?: string;
 /**
  * The creation date for this document as ISO 8601 date.
  */
 createdAt?: string;
 /**
  * The updated date for this document as ISO 8601 date.
  */
 updatedAt?: string;
}